% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R, R/functions_mmy.r
\name{cal.SPRetc3}
\alias{cal.SPRetc3}
\title{\enumerate{
\item Calculate spr, spr0, ypr by age ---
}}
\usage{
cal.SPRetc3(age, waa, maa, M, F, waa.mid, maa.mid, init.num = 1)

cal.SPRetc3(age, waa, maa, M, F, waa.mid, maa.mid, init.num = 1)
}
\description{
\enumerate{
\item Calculate spr, spr0, ypr by age ---
}

\enumerate{
\item Calculate spr, spr0, ypr by age ---
}
}
