% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R, R/functions_mmy.r
\name{calYPR.simple2}
\alias{calYPR.simple2}
\title{\enumerate{
\item Calculate spr, spr0, ypr, yield for the stock ---
}}
\usage{
calYPR.simple2(
  waa,
  maa,
  M,
  F,
  Fmulti,
  age,
  waa.mid,
  maa.mid,
  SR.coef = 1,
  is.ricker = F,
  steepness = 1
)

calYPR.simple2(
  waa,
  maa,
  M,
  F,
  Fmulti,
  age,
  waa.mid,
  maa.mid,
  SR.coef = 1,
  is.ricker = F,
  steepness = 1
)
}
\description{
\enumerate{
\item Calculate spr, spr0, ypr, yield for the stock ---
}

\enumerate{
\item Calculate spr, spr0, ypr, yield for the stock ---
}
}
