% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R, R/functions_mmy.r
\name{get.clark.data2}
\alias{get.clark.data2}
\title{\enumerate{
\item Further interace to get result of Clark (1991)
}}
\usage{
get.clark.data2(
  bpara,
  SR.coef = c(0.75, 0.875, 0.938, 1.386, 2.079, 2.773),
  is.ricker = c(F, F, F, T, T, T),
  f.vec = seq(from = 0, to = 1, by = 0.01),
  is.graph = F,
  steepness
)

get.clark.data2(
  bpara,
  SR.coef = c(0.75, 0.875, 0.938, 1.386, 2.079, 2.773),
  is.ricker = c(F, F, F, T, T, T),
  f.vec = seq(from = 0, to = 1, by = 0.01),
  is.graph = F,
  steepness
)
}
\description{
\enumerate{
\item Further interace to get result of Clark (1991)
}

\enumerate{
\item Further interace to get result of Clark (1991)
}
}
