% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R
\docType{data}
\name{Japan_dat}
\alias{Japan_dat}
\title{Read in functions ==============================================================================
Setting of runs ==============================================================================
\enumerate{
\item Read in data
}}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 12 columns.
}
\usage{
Japan_dat
}
\description{
Read in functions ==============================================================================
Setting of runs ==============================================================================
\enumerate{
\item Read in data
}
}
\keyword{datasets}
