% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R
\docType{data}
\name{output_folder}
\alias{output_folder}
\title{\enumerate{
\item set name of output file for Tables
}}
\format{
An object of class \code{character} of length 1.
}
\usage{
output_folder
}
\description{
\enumerate{
\item set name of output file for Tables
}
}
\keyword{datasets}
