% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R
\docType{data}
\name{set_analysis}
\alias{set_analysis}
\title{\enumerate{
\item Specify scenario
}}
\format{
An object of class \code{character} of length 1.
}
\usage{
set_analysis
}
\description{
\enumerate{
\item Specify scenario
}
}
\keyword{datasets}
