% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R
\docType{data}
\name{set_group_sr}
\alias{set_group_sr}
\title{\enumerate{
\item for scenario "group" and "species", specify to use only BH or both (BH and RI). if 1= only BH, 2= both
}}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
set_group_sr
}
\description{
\enumerate{
\item for scenario "group" and "species", specify to use only BH or both (BH and RI). if 1= only BH, 2= both
}
}
\keyword{datasets}
