% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R
\docType{data}
\name{set_hsd}
\alias{set_hsd}
\title{\enumerate{
\item Specify which quantile to use for steepness (h)
}}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
set_hsd
}
\description{
\enumerate{
\item Specify which quantile to use for steepness (h)
}
}
\keyword{datasets}
