% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_mmy.R
\docType{data}
\name{Stock_name}
\alias{Stock_name}
\title{End of settings ==============================================================================
Below is code (no need to change)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 12 columns.
}
\usage{
Stock_name
}
\description{
End of settings ==============================================================================
Below is code (no need to change)
}
\keyword{datasets}
